Remove all "local declaration hides ..." scope warnings.
Rationale:
No-one should be adding more variables with the same names as existing
ones in any scope.  These warnings are suppressed at the global scope so
that the existing warnings of this nature (which include cases where the
bug is required for backwards compatibility is concerned) can be left in
peace until such a time as they are revisited as part of a later
refactor.
This commit also contains removal of "if expression is constant"
warnings; these are for SDL_ASSERT and cannot be removed as a result.